sponza scene with shadows

+++ ShadowCore/Object/Object.cpp
@@ -3,7 +3,7 @@
 
 using namespace SC;
 
-uint64_t Object::Last_id = 1;
+uint32_t Object::Last_id = 1;
 
 Object::Object() {
 	enginePtr = Core::Engine;
